pyproject.toml
[project]
name = "my-video-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "av>=14.4.0",
    "customtkinter>=5.2.2",
    "pillow>=10.4.0",
    "python-vlc>=3.0.21203",
    "pytube>=15.0.0",
    "tkvideo>=0.1",
    "tkvideoplayer>=2.3",
    "ttkbootstrap>=1.13.8",
    "yt-dlp>=2025.5.22",
]
